fix: some CR comments
