+++ src/mito2/src/lib.rs
@@ -1,4 +1,3 @@
-#![feature(let_chains)]
 // Copyright 2023 Greptime Team
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -16,6 +15,8 @@
 //! # Mito
 //!
 //! Mito is the a region engine to store timeseries data.
+
+#![feature(let_chains)]
 
 #[cfg(any(test, feature = "test"))]
 pub mod test_util;
